+++ src/components/APISettingTree/item_dropdown.tsx
@@ -69,7 +69,7 @@ export default defineComponent({
 
     const addHTTPSetting = inject(
       addHTTPSettingKey,
-      addHTTPSettingDefaultValue
+      addHTTPSettingDefaultValue,
     );
     const addFolder = inject(addFolderKey, addFolderDefaultValue);
 
@@ -91,12 +91,16 @@ export default defineComponent({
         const apiSettings: APISetting[] = [];
         const folders: APIFolder[] = [];
         const name = apiFolderStore.findByID(id).name;
+        const folderIdList:string[] = [];
         const appendChildren = (folderId: string) => {
+          if (folderIdList.includes(folderId)) {
+            return;
+          }
+          folderIdList.push(folderId);
           const folder = apiFolderStore.findByID(folderId);
           if (!folder) {
             return;
           }
-          folders.push(folder);
           folder.children.split(",").forEach((child) => {
             const apiSetting = apiSettingStore.findByID(child);
             if (apiSetting) {
@@ -133,7 +137,7 @@ export default defineComponent({
           {
             const content = i18nCollection("deleteSettingContent").replace(
               "%s",
-              name
+              name,
             );
             const d = dialog.warning({
               title: i18nCollection("deleteSetting"),
@@ -205,7 +209,7 @@ export default defineComponent({
                 return convertRequestToCURL(
                   collection,
                   req,
-                  cookieStore.cookies
+                  cookieStore.cookies,
                 );
               })
               .then(writeTextToClipboard)
@@ -286,7 +290,7 @@ export default defineComponent({
               <FolderOpenOutline />
             </NIcon>
           ),
-        }
+        },
       );
       options.push(
         {
@@ -306,7 +310,7 @@ export default defineComponent({
               <DownloadOutline class="rotate270" />
             </NIcon>
           ),
-        }
+        },
       );
     } else {
       options.push(
@@ -337,7 +341,7 @@ export default defineComponent({
               <CopyOutline />
             </NIcon>
           ),
-        }
+        },
       );
     }
     options.push(
@@ -353,7 +357,7 @@ export default defineComponent({
             <TrashOutline />
           </NIcon>
         ),
-      }
+      },
     );
     return (
       <NDropdown